Add rust example using Axum framework and Tokio runtime
The output is constructed using html macro, which generates Virtual
Nodes and replicates "jsx" – style. Output is also compressed into gzip
to match with other rendering comparisons.

Result cases are intentionally ignored due to the nature of this comparison
example. We can always assume that we end in the happy path, and if not it's ok
to panic.

The crate has all optimizations enabled in dev profile (`opt-level = 3`).

The players data is slurped into memory of the application. The virtual
nodes are created on-demand when a request to root path `/` is received.
The app runs on port 8080.

Add ignore cases directly related to the rust-example like
`rust-axum/target` and `rust-axum/assets`.